Ajusto tipos de datos y pruebo consumo endpoint

+++ src/main/java/com/kometsales/inventory/dto/ProductCompanyItemDTO.java
@@ -2,11 +2,9 @@ package com.kometsales.inventory.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class ProductCompanyItemDTO {
     private String productName;
     private String basePrice;
-    private BigDecimal finalFreight;
+    private double finalFreight;
 }
