+++ src/main/java/com/kometsales/inventory/entity/ProductEntity.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -20,11 +19,11 @@ public class ProductEntity implements Serializable {
     @Id
     @GeneratedValue(strategy = IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private Integer id;
+    private int id;
     @Column(name = "name", length = 80)
     private String name;
     @Column(name = "freshCutValue", nullable = false, precision = 5)
-    private BigDecimal freshCutValue;
+    private double freshCutValue;
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "productEntity")
     @JsonIgnore
     private Set<InventoryEntity> inventoryEntities = new HashSet<>(0);
